Fix Core network wire widths in source

+++ Assets/Leviathan/Content/Scripts/CoreNetwork.cs
@@ -117,19 +117,10 @@ using UnityEngine;
 /// ---------------------------------------------------------------------------
 /// SIZE
 /// ---------------------------------------------------------------------------
-/// A typical native PlayerShipState is 44 bytes (41 fixed plus three empty
-/// array counts). Header is 5 bytes. A dynamic block with one active skill slot
-/// is about 8; an idle packet carries a one-byte zero count instead. The spec
-/// block for the current node set is about 27. Worst observed case is therefore
-/// roughly 84 bytes on the packets that carry a spec block and about 57 on the
-/// rest, dropping to 50 when no skill is active.
-///
-/// LiteNetLib does not fragment DeliveryMethod.Sequenced; NetPeer.Send throws
-/// TooBigPacketException past MTU minus header, and LiteNetLibTransport.Send has
-/// no try/catch. The native packet is variable length (auto targets, tractor
-/// containers, grapple links) with a worst case near 1044 bytes, so this class
-/// refuses to append anything once the native packet is already large. See
-/// MaxCombinedBytes.
+/// The Core header is 6 bytes. Payload size is bounded by MaxPayloadBytes and
+/// the finished native + Core packet is additionally bounded by MaxCombinedBytes.
+/// Dynamic presentation therefore degrades by omission on an unusually large
+/// native packet instead of risking LiteNetLib's non-fragmenting Sequenced path.
 ///
 /// ---------------------------------------------------------------------------
 /// FAILURE POLICY
@@ -179,7 +170,7 @@ public static class CoreNetwork
     public const int MaxClassNodes = 512;
     private const int MaxPackedSpecBytes = MaxClassNodes / 2;
 
-    /// <summary>schemaHash (2) + contentHash (4) + nodeCount (1).</summary>
+    /// <summary>classId (1) + schemaHash (2) + contentHash (4) + nodeCount (2).</summary>
     private const int SpecBlockHeaderBytes = 9;
 
     // Skill slot ids for the dynamic block. Stable forever once shipped; a
@@ -437,7 +428,6 @@ public static class CoreNetwork
     private static readonly Dictionary<MsgDamageEvent, CoreCombat.CombatEventMetadata>
         combatEventMetadata =
             new Dictionary<MsgDamageEvent, CoreCombat.CombatEventMetadata>(256);
-
     private static readonly Dictionary<MsgDamageResult, CoreCombat.CombatResultMetadata>
         combatResultMetadata =
             new Dictionary<MsgDamageResult, CoreCombat.CombatResultMetadata>(256);
@@ -569,7 +559,7 @@ public static class CoreNetwork
             writer.Write(CombatTrailerMagic);
             writer.Write(CombatProtocolVersion);
             writer.Write(CombatTrailerEvent);
-            writer.Write((ushort)payloadLength);
+            writer.Write((byte)payloadLength);
         }
         catch (Exception ex)
         {
@@ -635,7 +625,7 @@ public static class CoreNetwork
             writer.Write(CombatTrailerMagic);
             writer.Write(CombatProtocolVersion);
             writer.Write(CombatTrailerResult);
-            writer.Write((ushort)payloadLength);
+            writer.Write((byte)payloadLength);
         }
         catch (Exception ex)
         {
@@ -1488,7 +1478,7 @@ public static class CoreNetwork
             writer.Write(Magic1);
             writer.Write(ProtocolVersion);
             writer.Write(blockFlags);
-            writer.Write((byte)payloadLength);
+            writer.Write((ushort)payloadLength);
             writer.Write(scratch, 0, payloadLength);
 
             // Only count the burst down against packets that actually carried
